# Conflicts:
#	src/main/kotlin/audio/audioInput/AudioInput.kt
#	src/main/kotlin/lightOrgan/LightOrgan.kt
#	src/main/kotlin/lightOrgan/input/AudioInputManager.kt
#	src/main/kotlin/utilities/coroutines/SequencedFlow.kt
#	src/test/kotlin/audio/audioInput/AudioInputFixture.kt
#	src/test/kotlin/audio/audioInput/AudioInputTests.kt
#	src/test/kotlin/lightOrgan/LightOrganTests.kt
#	src/test/kotlin/lightOrgan/input/AudioInputManagerFixture.kt
#	src/test/kotlin/lightOrgan/input/AudioInputManagerTests.kt

+++ src/test/kotlin/audio/audioInput/AudioInputTests.kt
@@ -111,7 +111,7 @@ class AudioInputTests {
         assertEquals(readResults.size, received.size)
 
         received.forEachIndexed { index, sequenced ->
-            assertEquals(index, sequenced.sequenceNumber)
+            assertEquals(index.toLong(), sequenced.sequenceNumber)
             assertEquals(normalizedSamples, sequenced.value.samples)
             assertEquals(sut.format, sequenced.value.format)
         }
